% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{expr_demo}
\alias{expr_demo}
\title{Demo expression matrix (genes x samples, 500 x 24)}
\format{
Numeric matrix with genes as rows and samples as columns.
}
\usage{
expr_demo
}
\description{
Demo expression matrix (genes x samples, 500 x 24)
}
\examples{
data(expr_demo); dim(expr_demo)
}
\keyword{datasets}
