% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{sample_info_demo}
\alias{sample_info_demo}
\title{Demo sample metadata (24 samples)}
\format{
Data frame with columns: sample, group.
}
\usage{
sample_info_demo
}
\description{
Demo sample metadata (24 samples)
}
\examples{
data(sample_info_demo); head(sample_info_demo)
}
\keyword{datasets}
