% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermo_ssgsea.R
\name{thermo_ssgsea}
\alias{thermo_ssgsea}
\title{Immune and stroma signatures ssGSEAscores}
\usage{
thermo_ssgsea(
  expr,
  gene_sets,
  rank_method = c("max", "min"),
  na_to_bottom = TRUE
)
}
\arguments{
\item{expr}{numeric matrix with rows = genes, cols = samples}

\item{gene_sets}{named list: each element is a character vector of genes}

\item{rank_method}{one of c("max","min") for tie handling}

\item{na_to_bottom}{if TRUE, NAs are assigned the smallest rank (0)}
}
\value{
data.frame with rows = samples and columns = gene set names
}
\description{
ranks = data.T.rank(method=rank_method, na_option='bottom')
score = (sranks\strong{1.25).sum()/(sranks}0.25).sum() - (N - |S| + 1)/2
}
\examples{
set.seed(1)
expr <- matrix(runif(100), nrow = 10,
               dimnames = list(paste0("G",1:10), paste0("S",1:10)))
gsets <- list(SetA = paste0("G", c(1,3,5,7)), SetB = paste0("G", c(2,4,6)))
thermo_ssgsea(expr, gsets)
}
